Attempt to get lambda in a workable state
Instead of each function in math_operators.js simply concatenating the arguments, each operator iterates over the arguments and checks if it is an object; if so, it calls translate.ast_to_js on the object.

Circular imports need to be checked and simplified. Currently, lambda in defined inside translate, and it can likely be moved to lambda.js without import problems.